requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
charset-normalizer==3.1.0
discord.py==2.2.2
frozenlist==1.3.3
idna==3.4
lxml==4.9.2
multidict==6.0.4
mysql-connector-python==8.0.33
protobuf==3.20.3
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-decouple==3.8
six==1.16.0
soupsieve==2.4.1
yarl==1.8.2
